Working WS API + moved to separate crate
