+++ clients/native/examples/go-examples/websocket/binarysend.go
