+++ clients/native/examples/python-examples/websocket/binarysend.py
