+++ clients/native/src/websocket/mod.rs
@@ -12,7 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-pub mod api;
 pub(crate) mod handler;
 pub(crate) mod listener;
