+++ clients/native/websocket-requests/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "websocket-requests"
+version = "0.1.0"
+authors = ["Jędrzej Stuczyński <andrew@nymtech.net>"]
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+serde = { version = "1.0", features = ["derive"] }
+serde_json = "1.0"
+
+nymsphinx = { path = "../../../common/nymsphinx" }
