+++ clients/native/websocket-requests/src/error.rs
